Refactor home page greeting
Extract the greeting component into its own file.
